+++ src/main/kotlin/com/cmgapps/license/LicensesExtension.kt
@@ -19,18 +19,18 @@ package com.cmgapps.license
 import com.cmgapps.license.reporter.LicensesReportsContainer
 import groovy.lang.Closure
 import org.gradle.api.Action
-import org.gradle.util.ClosureBackedAction
+import org.gradle.util.ConfigureUtil
 
-open class LicensesExtension() {
+abstract class LicensesExtension {
     var reports: Action<in LicensesReportsContainer> = Action { }
         private set
 
     fun reports(closure: Closure<LicensesReportsContainer>) {
-        reports(ClosureBackedAction(closure))
+        reports = ConfigureUtil.configureUsing(closure)
     }
 
-    fun reports(configureAction: Action<in LicensesReportsContainer>) {
-        reports = configureAction
+    fun reports(action: Action<in LicensesReportsContainer>) {
+        reports = action
     }
 
     var additionalProjects = emptySet<String>()