+++ CHANGELOG.md
@@ -14,6 +14,39 @@
 
 ### Security
 
+## [3.2.0]
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+## [3.2.0]
+
+### Added
+
+- set task outputs to report files
+- add DSL to configure reports
+
+### Changed
+
+- use library to create CSV report
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
 ## [3.1.0]
 
 ### Added